requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
celery-with-redis==3.0
cffi==1.14.1
cryptography==3.0
Django==2.2.14
django-redis==4.10.0
importlib-metadata==1.7.0
kombu==4.6.11
mysql==0.0.2
mysqlclient==2.0.1
Pillow==7.2.0
pycparser==2.20
pycryptodomex==3.7.2
pyOpenSSL==19.1.0
pytest-runner==5.2
python-alipay-sdk==1.10.0
pytz==2020.1
redis==3.2.0
six==1.15.0
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
